requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.0
PyPDF2==3.0.1
jinja2==3.1.2
aiofiles
chromadb>=0.4.0
requests
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.10
